+++ src/main/java/net/accelbyte/sdk/api/eventlog/operations/event/GetEventByEventTypeAndEventIDHandler.java
@@ -64,9 +64,9 @@ public class GetEventByEventTypeAndEventIDHandler extends Operation {
     private Float eventId;
     private Float eventType;
     private String namespace;
-    private Float offset;
+    private Integer offset;
     private String endDate;
-    private Float pageSize;
+    private Integer pageSize;
     private String startDate;
 
     /**
@@ -82,9 +82,9 @@ public class GetEventByEventTypeAndEventIDHandler extends Operation {
             Float eventId,
             Float eventType,
             String namespace,
-            Float offset,
+            Integer offset,
             String endDate,
-            Float pageSize,
+            Integer pageSize,
             String startDate
     )
     {
