+++ src/main/java/net/accelbyte/sdk/api/eventlog/operations/event/GetEventByUserIDHandler.java
@@ -63,9 +63,9 @@ public class GetEventByUserIDHandler extends Operation {
      */
     private String namespace;
     private String userId;
-    private Float offset;
+    private Integer offset;
     private String endDate;
-    private Float pageSize;
+    private Integer pageSize;
     private String startDate;
 
     /**
@@ -79,9 +79,9 @@ public class GetEventByUserIDHandler extends Operation {
     public GetEventByUserIDHandler(
             String namespace,
             String userId,
-            Float offset,
+            Integer offset,
             String endDate,
-            Float pageSize,
+            Integer pageSize,
             String startDate
     )
     {
