+++ src/main/java/net/accelbyte/sdk/api/eventlog/operations/event_v2/GetEventSpecificUserV2Handler.java
@@ -62,8 +62,8 @@ public class GetEventSpecificUserV2Handler extends Operation {
     private String userId;
     private String endDate;
     private String eventName;
-    private Float offset;
-    private Float pageSize;
+    private Integer offset;
+    private Integer pageSize;
     private String startDate;
 
     /**
@@ -76,8 +76,8 @@ public class GetEventSpecificUserV2Handler extends Operation {
             String userId,
             String endDate,
             String eventName,
-            Float offset,
-            Float pageSize,
+            Integer offset,
+            Integer pageSize,
             String startDate
     )
     {
