+++ src/main/java/net/accelbyte/sdk/api/matchmaking/operations/matchmaking/SearchSessionsV2.java
@@ -69,8 +69,8 @@ public class SearchSessionsV2 extends Operation {
     private String matchID;
     private String partyID;
     private String userID;
-    private Float limit;
-    private Float offset;
+    private Integer limit;
+    private Integer offset;
 
     /**
     * @param namespace required
@@ -85,8 +85,8 @@ public class SearchSessionsV2 extends Operation {
             String matchID,
             String partyID,
             String userID,
-            Float limit,
-            Float offset
+            Integer limit,
+            Integer offset
     )
     {
         this.namespace = namespace;
